requirements_dev.txt
beautifulsoup4>=4.12.3
docutils>=0.20.1
GitPython>=3.1.42
graypy>=2.1.0
Jinja2>=3.1.3
numpy>=1.26.4
pandas>=2.2.1
pytest>=8.0.2
Requests>=2.31.0
setuptools>=68.2.2
Shapely>=2.0.3
sphinx_multiproject>=1.0.0rc1
urllib3>=2.2.1
tqdm
pyufunc

pillow
opencv-python
